feat: add ES2015/ES6 modules distribution in './module' folder
The files in './lib' are transpiled with 'babel-preset-es2015', which
by default uses the setting 'commonjs' for 'modules' (see: ./.babelrc).
This means that 'export something' declarations (ES2015/ES6 format)
are transpiled to 'exports.something = someotherthing' (CommonJS format),
making unused code elimination (with techniques such as 'tree shaking')
less effective or not possible.

The files in './module' are transpiled with 'babel-preset-es2015', but
'modules' is set to 'false' (see: ./.babelrc), so that 'export something'
declarations remain untouched. The 'module' key in 'package.json' tells
Webpack 2, Rollup and other ES6-aware bundlers that ES2015/ES6 modules
files are available and they should be used, ignoring the legacy format
files in './lib'.

Some references on this topic:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/import
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/export
http://jsmodules.io/
https://github.com/rollup/rollup/wiki/pkg.module