chore: add 'testing' BABEL_ENV and babel-polyfill for tests
This allows us to use neat methods such as ['lorem', 'ipsum'].includes('something'),
[Object.values](http://node.green/) and [Object.entries](http://node.green/) in
tests. By running 'tape -r babel-register -r babel-polyfill' in './package.json',
we tell Node to transpile the ES2015+ code to ES5 with 'register', and to install
global polyfills for  ES2015+ methods and instance methods that may not be supported
by a specific Node  version (4, 6, 7). For instance, if we didn't use 'babel-polyfill'
nor the node '--harmony' flag, Object.values/entries would cause an 'is not a function'
error in Node < 7.3.0 and [[...].includes](http://node.green/#Array-prototype-includes)
would cause the same error in Node < 6.9.3.

References:
- [Official Babel Polyfill documentation](https://babeljs.io/docs/usage/polyfill/)
- [Node.js ES2015 support tables](http://node.green/)
- [Object.values](https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_objects/Object/values) on MDN
- [Object.entries](https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_objects/Object/entries) on MDN
- [Array.prototype.includes](https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_Objects/Array/includes) on MDN