chore(flow): flow 0.53+ compatibility - 'FlagIcon' types
- use the new 'StatelessFunctionalComponent' Flow type instead of
manually typing with React.element.
- make the Flow type 'FlagIconPropsType' non-polymorphic, as users who
need 'customCodes' have to use 'CustomFlagIconFactory' at the moment,
and this Flow version doesn't seem to accept the previous 'T -> void'
workaround.
- Ideally there would just be one 'FlagIconFactory' with optional support
for 'customCodes', but Flow-wise it doesn't seem to be easy ... so we
stick with this approach for now (PR for that - passing all Flow tests -
is welcome).

SEE: https://flow.org/en/docs/react/types/#toc-react-statelessfunctionalcomponent